pptx_extractor/requirements.txt
python-pptx==0.6.21
Pillow==10.2.0
python-docx==0.8.11
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
opencv-python==4.8.0
spacy==3.7.2
nltk==3.8.1
pandas==2.1.0
numpy==1.24.3 